fixed some memory leak, as calendar stopped working with too many calendars and to long entries.

+++ stm32_application/usart.h
@@ -9,7 +9,7 @@ void Serial_USART_Initialize();
 void USART_StartDMA();
 void USART_Write(const char * txt);
 void USART_WriteInt(int v);
-uint32_t USART_ReadString(char *data, uint16_t max_size);
+uint32_t USART_ReadString(char *data, uint32_t max_size);
 void USART_WriteStatus (uint8_t status);
 unsigned char USART_ReadByteSync(USART_TypeDef *USARTx, unsigned * waiter);
